feat: updated build and types
