--- src/react/index.ts
@@ -1,3 +1,0 @@
-import {useController} from './extension-load-queue';
-
-export {useController};
